build: Treat warnings as errors (#923)
* squash and rebase andronov-alexey:treat_warnings_as_errors from pull request #850
* fix remaining warnings in java lib

+++ src/lib/app/Application.cpp
@@ -450,7 +450,7 @@ void Application::updateRecentProjects(const FilePath& projectSettingsFilePath)
 		}
 
 		recentProjects.insert(recentProjects.begin(), projectSettingsFilePath);
-		while (recentProjects.size() > appSettings->getMaxRecentProjectsCount())
+		while (static_cast<int>(recentProjects.size()) > appSettings->getMaxRecentProjectsCount())
 		{
 			recentProjects.pop_back();
 		}
