+++ src/lib/data/parser/TaskParseWrapper.cpp
@@ -36,7 +36,7 @@ Task::TaskState TaskParseWrapper::doUpdate(std::shared_ptr<Blackboard> blackboar
 
 void TaskParseWrapper::doExit(std::shared_ptr<Blackboard> blackboard)
 {
-	float duration = TimeStamp::durationSeconds(m_start);
+	float duration = static_cast<float>(TimeStamp::durationSeconds(m_start));
 	blackboard->update<float>(
 		"index_time", [duration](float currentDuration) { return currentDuration + duration; });
 }
