+++ src/lib_gui/qt/project_wizard/content/QtProjectWizardContentCStandard.cpp
@@ -33,7 +33,7 @@ void QtProjectWizardContentCStandard::load()
 		std::vector<std::wstring> standards = m_sourceGroupSettings->getAvailableCStandards();
 		for (size_t i = 0; i < standards.size(); i++)
 		{
-			m_standard->insertItem(i, QString::fromStdWString(standards[i]));
+			m_standard->insertItem(static_cast<int>(i), QString::fromStdWString(standards[i]));
 		}
 
 		m_standard->setCurrentText(QString::fromStdWString(m_sourceGroupSettings->getCStandard()));
