+++ src/test/SourceGroupTestSuite.cpp
@@ -240,7 +240,7 @@ void generateAndCompareExpectedOutput(
 			expectedOutput->getLineCount() == output->getLineCount());
 		if (expectedOutput->getLineCount() == output->getLineCount())
 		{
-			for (size_t i = 1; i <= expectedOutput->getLineCount(); i++)
+			for (unsigned int i = 1; i <= expectedOutput->getLineCount(); i++)
 			{
 				REQUIRE(expectedOutput->getLine(i) == output->getLine(i));
 			}
